chore: rename user-auth to strategy.ts and make validate return AuthUserDto
