+++ server/apps/immich/src/modules/immich-auth/guards/auth.guard.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { AuthGuard as PassportAuthGuard } from '@nestjs/passport';
 import { API_KEY_STRATEGY } from '../strategies/api-key.strategy';
-import { AUTH_COOKIE_STRATEGY } from '../strategies/user-auth-strategy.service';
+import { AUTH_COOKIE_STRATEGY } from '../strategies/user-auth.strategy';
 import { PUBLIC_SHARE_STRATEGY } from '../strategies/public-share.strategy';
 
 @Injectable()
